requirements.txt
python-dotenv
agentmemory
agentbrowser
easycompletion
agentaction
agentloop
pyfiglet
rich